+++ src/main/java/com/vuvk/voxelspace/Global.java
@@ -24,29 +24,29 @@ public final class Global {
     public final static int THREADS_COUNT = Runtime.getRuntime().availableProcessors() + 1;
     
     public static boolean fogEnabled = true;
-    public static double fogStart = 10.0;
-    public static double fogEnd   = 100.0;
+    public static float fogStart = 10.0f;
+    public static float fogEnd   = 100.0f;
     public static int fogColor = 0xFFCCCCFF;
     public static int skyColor = 0xFF9999FF;
-    public static double detalization = 0.0025;
+    public static float detalization = 0.0025f;
     public static boolean multithreading = true;
-    public static double deltaTime = 0;    
+    public static float deltaTime = 0;    
     public static long curTick;
     public static long lastTick;
     public static int fps = 0;
     private static int frames = 0;
-    private static double fpsDelay;
+    private static float fpsDelay;
     
     public static void update() {
         curTick = System.currentTimeMillis();
         
-        deltaTime = (curTick - lastTick) / 1000.0;
+        deltaTime = (curTick - lastTick) / 1000.0f;
         
         if (fpsDelay < 1.0) {
             fpsDelay += deltaTime;
             ++frames;
         } else {
-            fpsDelay = 0.0;
+            fpsDelay = 0.0f;
             fps = frames;
             frames = 0;
         }
